+++ include/Ark/Compiler/Common.hpp
@@ -101,6 +101,8 @@ namespace Ark::internal
         "del"
     };
 
+    constexpr std::string_view HiddenSymbolSuffix = "#hidden";
+
     namespace Language
     {
         constexpr std::string_view AppendInPlace = "append!";